% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC Curve}
\usage{
plot_roc(actual, prob)
}
\arguments{
\item{actual}{Vector of actual binary outcomes}

\item{prob}{Vector of predicted probabilities}
}
\value{
A ggplot object showing ROC curve
}
\description{
Plot ROC Curve
}
