% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.nr_logistic}
\alias{predict.nr_logistic}
\title{Predict method for nr_logistic objects}
\usage{
\method{predict}{nr_logistic}(object, newdata, type = c("response", "class"), threshold = 0.5, ...)
}
\arguments{
\item{object}{An object of class "nr_logistic"}

\item{newdata}{A data frame containing new predictor values}

\item{type}{Type of prediction: "response" (default, gives probabilities) or "class" (gives 0/1)}

\item{threshold}{Threshold for classification when type = "class" (default = 0.5)}

\item{...}{Additional arguments (ignored)}
}
\value{
A vector of predicted probabilities or classes
}
\description{
Predict method for nr_logistic objects
}
